+++ trunk/src/ecosim/model/architecture/functions/LinearFunction.java
@@ -2,15 +2,15 @@ package ecosim.model.architecture.functions;
 
 public class LinearFunction implements ArchitectureFunction {
 
-	private final float M;
-	private final float B;
+	private final double M;
+	private final double B;
 	
-	public LinearFunction(float m, float b) {
+	public LinearFunction(double m, double b) {
 		this.M = m;
 		this.B = b;
 	}
 	@Override
-	public float evaluate(float x) {
+	public double evaluate(double x) {
 		return M * x + B;
 	}
